requirements.txt
Django>=4.0,<5.0
djangorestframework==3.15.2
psycopg2==2.9.10
django-environ
django-redis